Move chat to it's own cog, added info_chat
