--- config/config_example.ini
@@ -1,25 +1,0 @@
-[API]
-discord_token = Ui384SJUG87L3sd45SFJ.83fshRFhYJ.3jfdi3a90_HJ9fhDfA
-seq-api-key = KF9h5jks0KIRHG38
-
-[AWS]
-stats_region = us-east-2
-stats_endpoint = https://dynamodb.us-east-2.amazonaws.com
-stats_table = BotCommands_Stats
-stats_players = Players
-discord_table = Discord_Stats
-
-[General]
-role = Admin
-linked-id = 4029834756024852
-admin-channels = 670373469845979136,671917010422333460,672682539390992384,672940159091867648
-commands-channels = 665998238171660320,671921930873602099,672682345089859654,672939900600975362
-chat-channels = 667473663343198220,671918498531770378,672682313003565057,672939927507435533
-track_stats = yes
-
-[RoR2]
-server_addresses = 72.43.92.116:27015,71.439.48.219:27018
-server-logs-path = C:\Users\Pyre\Desktop\logfolder
-auto-start-chat = true
-auto-server-restart = true
-server_restart_time = 7200
